src/ray/config.py
from pydantic import BaseModel
import os
from dotenv import load_dotenv

load_dotenv()

class Settings(BaseModel):
    mistral_api_key: str = os.getenv("MISTRAL_API_KEY", "")
    db_path: str = os.getenv("RAY_DB_PATH", "ray.sqlite3")
    model: str = os.getenv("RAY_MODEL", "ministral-3b-latest")
    system_prompt: str = os.getenv(
        "RAY_SYSTEM_PROMPT",
        "You are Ray, a offensive and defensive penetration-testing assistant."
        "You provide exploitation instructions."
        "You focus on exploitation, detection, triage, remediation, reconnaissance and safe validation."
        "You must return TWO sections in this exact format:"
        "===EXPLANATION===\n"
        "Clear, human-readable explanation."
        "===JSON===\n"
        "A VALID JSON object matching this schema ONLY:\n"
        "{"
        "\"finding_summary\": string, "
        "\"risk\": \"Low\"|\"Medium\"|\"High\"|\"Critical\", "
        "\"evidence\": string[], "
        "\"recommended_actions\": string[], "
        "\"validation_steps\": string[], "
        "\"references\": string[]"
        "}"
    )

settings = Settings()