pyproject.toml
[project]
name = "ray"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.0",
  "python-dotenv>=1.0.0",
  "mistralai>=1.0.0",
  "jsonschema>=4.21.1",
  "pydantic>=2.7.0"
]

[project.scripts]
ray = "ray.cli:app"